New project structure with last react and deno versions. Added docker compose and dockerfiles
