+++ api/deps.ts
@@ -1,0 +1,10 @@
+// A place to store all dependent URLs
+// place standar library modules bellow
+
+
+//TODO: export {} from std/dasdsa/mod.ts
+
+//place third party modules bellow
+export {Application, Router} from "third/oak@v10.5.1/mod.ts";
+
+//TODO: export {} from third/dasd@version/mod.ts
