--- deps.ts
@@ -1,11 +1,0 @@
-// A place to store all dependent URLs
-// place standar library modules bellow
-
-
-//TODO: export {} from std/dasdsa/mod.ts
-
-//place third party modules bellow
-// @ts-ignore
-export {Application, Router} from "third/oak@v10.5.1/mod.ts";
-
-//TODO: export {} from third/dasd@version/mod.ts
