+++ frontend/Dockerfile
@@ -1,0 +1,9 @@
+FROM node:16-alpine3.15
+
+WORKDIR /usr/app
+
+COPY package.json ./
+
+RUN npm install
+
+COPY . .
