--- import_map.json
@@ -1,6 +1,0 @@
-{
-    "imports":{
-        "std/": "https://deno.land/std@0.136.0/",
-        "third/": "https://deno.land/x/"
-    }
-}
