--- view/App.tsx
@@ -1,27 +1,0 @@
-import {React} from './deps.ts';
-
-function App() {
-  return (
-    <>
-      <link rel={'stylesheet'} href={'./static/css/App.css'}/>
-      <div className="App">
-        <header className="App-header">
-          <img src={'./static/images/logo.svg'} className="App-logo" alt="logo"/>
-          <p>
-            Edit <code>App.tsx</code> and save to reload.
-          </p>
-          <a
-            className="App-link"
-            href="https://reactjs.org"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            Learn React
-          </a>
-        </header>
-      </div>
-    </>
-  );
-}
-
-export default App;
