--- view/index.tsx
@@ -1,9 +1,0 @@
-import {React, react, ReactDOM} from './deps.ts';
-import App from './App.tsx';
-
-ReactDOM.render(
-  <react.StrictMode>
-    <App />
-  </react.StrictMode>,
-  document.getElementById('root')
-);
