--- view/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "compilerOptions": {
-    "noImplicitAny": false,
-    "lib": [
-      "dom",
-      "dom.iterable",
-      "esnext"
-    ]
-  }
-}
